requirements.txt
requests
beautifulsoup4
feedparser
transformers
torch
python-dotenv
streamlit